Move CoreContentScopeScripts back to the impl module
